+++ ejercicios/src/main/java/U1/Tarea2/Ejercicio4.java
@@ -6,7 +6,7 @@ public class Ejercicio4 {
     public static void main(String[] args) {
         System.out.print("Introduzca la distancia en millas: ");
         Scanner teclado = new Scanner(System.in);
-        double millas = teclado.nextInt();
+        double millas = teclado.nextDouble();
         System.out.print("La equivalencia en metros es: ");
         System.out.println(millas * 1609);
     }
